// packages/backend.ai-ui/src/__generated__/BAIObjectStorageSelectQuery.graphql.ts
/**
 * @generated SignedSource<<83209a4a1e70df0b60d680039d630981>>
 * @lightSyntaxTransform
 * @nogrep
 */

/* tslint:disable */
/* eslint-disable */
// @ts-nocheck

import { ConcreteRequest } from 'relay-runtime';
export type BAIObjectStorageSelectQuery$variables = {
  limit: number;
  offset: number;
};
export type BAIObjectStorageSelectQuery$data = {
  readonly objectStorages: {
    readonly count: number;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly name: string;
      };
    }>;
  } | null | undefined;
};
export type BAIObjectStorageSelectQuery = {
  response: BAIObjectStorageSelectQuery$data;
  variables: BAIObjectStorageSelectQuery$variables;
};

const node: ConcreteRequest = (function(){
var v0 = {
  "defaultValue": null,
  "kind": "LocalArgument",
  "name": "limit"
},
v1 = {
  "defaultValue": null,
  "kind": "LocalArgument",
  "name": "offset"
},
v2 = [
  {
    "alias": null,
    "args": [
      {
        "kind": "Variable",
        "name": "limit",
        "variableName": "limit"
      },
      {
        "kind": "Variable",
        "name": "offset",
        "variableName": "offset"
      }
    ],
    "concreteType": "ObjectStorageConnection",
    "kind": "LinkedField",
    "name": "objectStorages",
    "plural": false,
    "selections": [
      {
        "alias": null,
        "args": null,
        "kind": "ScalarField",
        "name": "count",
        "storageKey": null
      },
      {
        "alias": null,
        "args": null,
        "concreteType": "ObjectStorageEdge",
        "kind": "LinkedField",
        "name": "edges",
        "plural": true,
        "selections": [
          {
            "alias": null,
            "args": null,
            "concreteType": "ObjectStorage",
            "kind": "LinkedField",
            "name": "node",
            "plural": false,
            "selections": [
              {
                "alias": null,
                "args": null,
                "kind": "ScalarField",
                "name": "id",
                "storageKey": null
              },
              {
                "alias": null,
                "args": null,
                "kind": "ScalarField",
                "name": "name",
                "storageKey": null
              }
            ],
            "storageKey": null
          }
        ],
        "storageKey": null
      }
    ],
    "storageKey": null
  }
];
return {
  "fragment": {
    "argumentDefinitions": [
      (v0/*: any*/),
      (v1/*: any*/)
    ],
    "kind": "Fragment",
    "metadata": null,
    "name": "BAIObjectStorageSelectQuery",
    "selections": (v2/*: any*/),
    "type": "Query",
    "abstractKey": null
  },
  "kind": "Request",
  "operation": {
    "argumentDefinitions": [
      (v1/*: any*/),
      (v0/*: any*/)
    ],
    "kind": "Operation",
    "name": "BAIObjectStorageSelectQuery",
    "selections": (v2/*: any*/)
  },
  "params": {
    "cacheID": "b5de4f28fbe16c4e489840b9f8fda154",
    "id": null,
    "metadata": {},
    "name": "BAIObjectStorageSelectQuery",
    "operationKind": "query",
    "text": "query BAIObjectStorageSelectQuery(\n  $offset: Int!\n  $limit: Int!\n) {\n  objectStorages(offset: $offset, limit: $limit) {\n    count\n    edges {\n      node {\n        id\n        name\n      }\n    }\n  }\n}\n"
  }
};
})();

(node as any).hash = "f7e30d72c24d41e1a8a141531a425420";

export default node;